TINKERPOP-1568 Changed when side-effects were set to child traversals in strategy application
We were doing this "before" strategy application but it caused trouble with strategies like TranslationStrategy which modified side-effects as part of their execution. That said, TranslationStrategy is a "test" strategy and I wonder if we'd actually see that kind of thing happening in a normal strategy implementation. Moving when side-effects were set not only fixed the problem but also made me question whether the parent strategy needed to be set to all the children. Removing that all together, did something interesting because it essentially removed an entire recursive iteration of the traversal structure.